보류(back-end)/requirements.txt
apache-beam==2.61.0
apache-flink==2.1.1
apache-flink-libraries==2.1.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.4.0
avro==1.12.1
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==2.2.1
crcmod==1.7
cryptography==45.0.5
dill==0.3.1.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==64.0.0
djangorestframework==3.16.0
dnspython==2.8.0
docopt==0.6.2
dotenv==0.9.9
fastavro==1.12.1
fasteners==0.20
feedparser==6.0.12
find_libpython==0.5.0
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
Naked==0.1.32
numpy==2.1.3
oauthlib==3.3.1
objsize==0.7.1
orjson==3.11.4
packaging==25.0
pandas==2.2.3
pemja==0.5.5
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.4
rpds-py==0.29.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
zstandard==0.25.0
